Move profiles CLI from picard/profiles/cli.py to picard/cli/profiles.py
All CLI subcommand modules now live under picard/cli/ for consistency
with the plugins CLI.
